move evaluation actions to separate file
